# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "LARAVEL Rest API DOCUMENTATION",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/dashboard/parkingLots/{parkingLotId}/{period}": {
            "get": {
                "tags": [
                    "DashBoard"
                ],
                "summary": "get statics of one parkingLot with a Id",
                "operationId": "parkingLotBookingStats",
                "parameters": [
                    {
                        "name": "parkingLotId",
                        "in": "path",
                        "description": "id parkingLot management",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1000000
                    },
                    {
                        "name": "period",
                        "in": "path",
                        "description": "period day or week or month or year",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "day"
                    }
                ],
                "responses": {
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/api/dashboard/parkingLots/{userId}": {
            "get": {
                "tags": [
                    "DashBoard"
                ],
                "summary": "get all parkinglot by user management",
                "operationId": "getParkingUserManage",
                "parameters": [
                    {
                        "name": "userId",
                        "in": "path",
                        "description": "id user management",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1000000
                    }
                ],
                "responses": {
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/api/dashboard/{userId}/revenue/{period}": {
            "get": {
                "tags": [
                    "DashBoard"
                ],
                "summary": "Statistics parking lot block, revenue,userBooking by number",
                "operationId": "getRevenueDetails",
                "parameters": [
                    {
                        "name": "userId",
                        "in": "path",
                        "description": "id user manage parking lto lot",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1000000
                    },
                    {
                        "name": "period",
                        "in": "path",
                        "description": "period day or week or month or year",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "day"
                    }
                ],
                "responses": {
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "passport": {
                "type": "apiKey",
                "description": "Laravel passport oauth2 security.",
                "in": "header",
                "name": "Authorization",
                "scheme": "https",
                "flows": {
                    "password": {
                        "authorizationUrl": "http://localhost/oauth/authorize",
                        "tokenUrl": "http://localhost/oauth/token",
                        "refreshUrl": "http://localhost/token/refresh",
                        "scopes": []
                    }
                }
            }
        }
    }
}